change system

+++ system/system_monitor/test/src/cpu_monitor/test_intel_cpu_monitor.cpp
@@ -1,21 +1,24 @@
-// Copyright 2020 Tier IV, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
+/*
+ * Copyright 2020 Autoware Foundation. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 
-#include "system_monitor/cpu_monitor/intel_cpu_monitor.hpp"
+#include <pthread.h>
 
-#include <msr_reader/msr_reader.hpp>
-#include <rclcpp/rclcpp.hpp>
+#include <fstream>
+#include <string>
+#include <vector>
 
 #include <boost/algorithm/string.hpp>
 #include <boost/archive/text_oarchive.hpp>
@@ -24,18 +27,16 @@
 
 #include <fmt/format.h>
 #include <gtest/gtest.h>
-#include <pthread.h>
+#include <ros/ros.h>
 
-#include <fstream>
-#include <memory>
-#include <string>
-#include <vector>
+#include <msr_reader/msr_reader.h>
+#include <system_monitor/cpu_monitor/intel_cpu_monitor.h>
 
 static constexpr const char * TEST_FILE = "test";
 static constexpr const char * DOCKER_ENV = "/.dockerenv";
 
 namespace fs = boost::filesystem;
-using DiagStatus = diagnostic_msgs::msg::DiagnosticStatus;
+using DiagStatus = diagnostic_msgs::DiagnosticStatus;
 
 char ** argv_;
 
@@ -44,29 +45,26 @@ class TestCPUMonitor : public CPUMonitor
   friend class CPUMonitorTestSuite;
 
 public:
-  TestCPUMonitor(const std::string & node_name, const rclcpp::NodeOptions & options)
-  : CPUMonitor(node_name, options)
-  {
-  }
+  TestCPUMonitor(const ros::NodeHandle & nh, const ros::NodeHandle & pnh) : CPUMonitor(nh, pnh) {}
 
-  void diagCallback(const diagnostic_msgs::msg::DiagnosticArray::ConstSharedPtr diag_msg)
+  void diagCallback(const diagnostic_msgs::DiagnosticArray::ConstPtr & diag_msg)
   {
     array_ = *diag_msg;
   }
 
   void addTempName(const std::string & path) { temps_.emplace_back(path, path); }
-  void clearTempNames() { temps_.clear(); }
-  bool isTempNamesEmpty() { return temps_.empty(); }
+  void clearTempNames(void) { temps_.clear(); }
+  bool isTempNamesEmpty(void) { return temps_.empty(); }
 
   void addFreqName(int index, const std::string & path) { freqs_.emplace_back(index, path); }
-  void clearFreqNames() { freqs_.clear(); }
+  void clearFreqNames(void) { freqs_.clear(); }
 
   void setMpstatExists(bool mpstat_exists) { mpstat_exists_ = mpstat_exists; }
 
   void changeUsageWarn(float usage_warn) { usage_warn_ = usage_warn; }
   void changeUsageError(float usage_error) { usage_error_ = usage_error; }
 
-  void update() { updater_.force_update(); }
+  void update(void) { updater_.force_update(); }
 
   const std::string removePrefix(const std::string & name)
   {
@@ -85,14 +83,14 @@ public:
   }
 
 private:
-  diagnostic_msgs::msg::DiagnosticArray array_;
-  const std::string prefix_ = std::string(this->get_name()) + ": ";
+  diagnostic_msgs::DiagnosticArray array_;
+  const std::string prefix_ = ros::this_node::getName().substr(1) + ": ";
 };
 
 class CPUMonitorTestSuite : public ::testing::Test
 {
 public:
-  CPUMonitorTestSuite()
+  CPUMonitorTestSuite() : nh_(""), pnh_("~")
   {
     // Get directory of executable
     const fs::path exe_path(argv_[0]);
@@ -102,43 +100,31 @@ public:
   }
 
 protected:
+  ros::NodeHandle nh_, pnh_;
   std::unique_ptr<TestCPUMonitor> monitor_;
-  rclcpp::Subscription<diagnostic_msgs::msg::DiagnosticArray>::SharedPtr sub_;
+  ros::Subscriber sub_;
   std::string exe_dir_;
   std::string mpstat_;
 
-  void SetUp()
+  void SetUp(void)
   {
-    using std::placeholders::_1;
-    rclcpp::init(0, nullptr);
-    rclcpp::NodeOptions node_options;
-    monitor_ = std::make_unique<TestCPUMonitor>("test_cpu_monitor", node_options);
-    sub_ = monitor_->create_subscription<diagnostic_msgs::msg::DiagnosticArray>(
-      "/diagnostics", 1000, std::bind(&TestCPUMonitor::diagCallback, monitor_.get(), _1));
+    monitor_ = std::make_unique<TestCPUMonitor>(nh_, pnh_);
+    sub_ = nh_.subscribe("/diagnostics", 1000, &TestCPUMonitor::diagCallback, monitor_.get());
     monitor_->getTempNames();
     monitor_->getFreqNames();
 
     // Remove test file if exists
-    if (fs::exists(TEST_FILE)) {
-      fs::remove(TEST_FILE);
-    }
+    if (fs::exists(TEST_FILE)) fs::remove(TEST_FILE);
     // Remove dummy executable if exists
-    if (fs::exists(mpstat_)) {
-      fs::remove(mpstat_);
-    }
+    if (fs::exists(mpstat_)) fs::remove(mpstat_);
   }
 
-  void TearDown()
+  void TearDown(void)
   {
     // Remove test file if exists
-    if (fs::exists(TEST_FILE)) {
-      fs::remove(TEST_FILE);
-    }
+    if (fs::exists(TEST_FILE)) fs::remove(TEST_FILE);
     // Remove dummy executable if exists
-    if (fs::exists(mpstat_)) {
-      fs::remove(mpstat_);
-    }
-    rclcpp::shutdown();
+    if (fs::exists(mpstat_)) fs::remove(mpstat_);
   }
 
   bool findValue(const DiagStatus status, const std::string & key, std::string & value)  // NOLINT
@@ -152,7 +138,7 @@ protected:
     return false;
   }
 
-  void modifyPath()
+  void modifyPath(void)
   {
     // Modify PATH temporarily
     auto env = boost::this_process::environment();
@@ -180,9 +166,7 @@ void * msr_reader(void * args)
 
   // Create a new socket
   int sock = socket(AF_INET, SOCK_STREAM, 0);
-  if (sock < 0) {
-    return nullptr;
-  }
+  if (sock < 0) return nullptr;
 
   // Allow address reuse
   int ret = 0;
@@ -253,9 +237,7 @@ void * msr_reader(void * args)
       // Wait for recv timeout
       while (true) {
         pthread_mutex_lock(&mutex);
-        if (stop_thread) {
-          break;
-        }
+        if (stop_thread) break;
         pthread_mutex_unlock(&mutex);
         sleep(1);
       }
@@ -285,9 +267,7 @@ void * msr_reader(void * args)
 TEST_F(CPUMonitorTestSuite, tempWarnTest)
 {
   // Skip test if process runs inside CI environment
-  if (monitor_->isTempNamesEmpty() && fs::exists(DOCKER_ENV)) {
-    return;
-  }
+  if (monitor_->isTempNamesEmpty() && fs::exists(DOCKER_ENV)) return;
 
   // Verify normal behavior
   {
@@ -295,8 +275,8 @@ TEST_F(CPUMonitorTestSuite, tempWarnTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
 
     // Verify
     DiagStatus status;
@@ -317,8 +297,8 @@ TEST_F(CPUMonitorTestSuite, tempWarnTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
 
     // Verify
     DiagStatus status;
@@ -336,8 +316,8 @@ TEST_F(CPUMonitorTestSuite, tempWarnTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
 
     // Verify
     DiagStatus status;
@@ -349,9 +329,7 @@ TEST_F(CPUMonitorTestSuite, tempWarnTest)
 TEST_F(CPUMonitorTestSuite, tempErrorTest)
 {
   // Skip test if process runs inside CI environment
-  if (monitor_->isTempNamesEmpty() && fs::exists(DOCKER_ENV)) {
-    return;
-  }
+  if (monitor_->isTempNamesEmpty() && fs::exists(DOCKER_ENV)) return;
 
   // Verify normal behavior
   {
@@ -359,8 +337,8 @@ TEST_F(CPUMonitorTestSuite, tempErrorTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
 
     // Verify
     DiagStatus status;
@@ -381,8 +359,9 @@ TEST_F(CPUMonitorTestSuite, tempErrorTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
+
     // Verify
     DiagStatus status;
     ASSERT_TRUE(monitor_->findDiagStatus("CPU Temperature", status));
@@ -399,8 +378,8 @@ TEST_F(CPUMonitorTestSuite, tempErrorTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
 
     // Verify
     DiagStatus status;
@@ -418,8 +397,8 @@ TEST_F(CPUMonitorTestSuite, tempTemperatureFilesNotFoundTest)
   monitor_->update();
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
 
   // Verify
   DiagStatus status;
@@ -437,8 +416,8 @@ TEST_F(CPUMonitorTestSuite, tempFileOpenErrorTest)
   monitor_->update();
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
 
   // Verify
   DiagStatus status;
@@ -458,8 +437,8 @@ TEST_F(CPUMonitorTestSuite, usageWarnTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
 
     // Verify
     DiagStatus status;
@@ -477,8 +456,8 @@ TEST_F(CPUMonitorTestSuite, usageWarnTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
 
     // Verify
     DiagStatus status;
@@ -495,8 +474,8 @@ TEST_F(CPUMonitorTestSuite, usageWarnTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
 
     // Verify
     DiagStatus status;
@@ -513,8 +492,8 @@ TEST_F(CPUMonitorTestSuite, usageErrorTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
 
     // Verify
     DiagStatus status;
@@ -532,8 +511,8 @@ TEST_F(CPUMonitorTestSuite, usageErrorTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
 
     // Verify
     DiagStatus status;
@@ -550,8 +529,8 @@ TEST_F(CPUMonitorTestSuite, usageErrorTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
 
     // Verify
     DiagStatus status;
@@ -569,8 +548,8 @@ TEST_F(CPUMonitorTestSuite, usageMpstatNotFoundTest)
   monitor_->update();
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
 
   // Verify
   DiagStatus status;
@@ -592,8 +571,8 @@ TEST_F(CPUMonitorTestSuite, load1WarnTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
 
     // Verify
     DiagStatus status;
@@ -611,8 +590,8 @@ TEST_F(CPUMonitorTestSuite, load5WarnTest)
     monitor_->update();
 
     // Give time to publish
-    rclcpp::WallRate(2).sleep();
-    rclcpp::spin_some(monitor_->get_node_base_interface());
+    ros::WallDuration(0.5).sleep();
+    ros::spinOnce();
 
     // Verify
     DiagStatus status;
@@ -629,7 +608,7 @@ TEST_F(CPUMonitorTestSuite, throttlingTest)
   ThreadTestMode mode = Normal;
   pthread_create(&th, nullptr, msr_reader, &mode);
   // Wait for thread started
-  rclcpp::WallRate(10).sleep();
+  ros::WallDuration(0.1).sleep();
 
   // Publish topic
   monitor_->update();
@@ -637,8 +616,8 @@ TEST_F(CPUMonitorTestSuite, throttlingTest)
   pthread_join(th, NULL);
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
 
   // Verify
   DiagStatus status;
@@ -652,7 +631,7 @@ TEST_F(CPUMonitorTestSuite, throttlingThrottlingTest)
   ThreadTestMode mode = Throttling;
   pthread_create(&th, nullptr, msr_reader, &mode);
   // Wait for thread started
-  rclcpp::WallRate(10).sleep();
+  ros::WallDuration(0.1).sleep();
 
   // Publish topic
   monitor_->update();
@@ -660,8 +639,8 @@ TEST_F(CPUMonitorTestSuite, throttlingThrottlingTest)
   pthread_join(th, NULL);
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
 
   // Verify
   DiagStatus status;
@@ -676,7 +655,7 @@ TEST_F(CPUMonitorTestSuite, throttlingReturnsErrorTest)
   ThreadTestMode mode = ReturnsError;
   pthread_create(&th, nullptr, msr_reader, &mode);
   // Wait for thread started
-  rclcpp::WallRate(10).sleep();
+  ros::WallDuration(0.1).sleep();
 
   // Publish topic
   monitor_->update();
@@ -684,8 +663,8 @@ TEST_F(CPUMonitorTestSuite, throttlingReturnsErrorTest)
   pthread_join(th, NULL);
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
 
   // Verify
   DiagStatus status;
@@ -703,7 +682,7 @@ TEST_F(CPUMonitorTestSuite, throttlingRecvTimeoutTest)
   ThreadTestMode mode = RecvTimeout;
   pthread_create(&th, nullptr, msr_reader, &mode);
   // Wait for thread started
-  rclcpp::WallRate(10).sleep();
+  ros::WallDuration(0.1).sleep();
 
   // Publish topic
   monitor_->update();
@@ -715,8 +694,8 @@ TEST_F(CPUMonitorTestSuite, throttlingRecvTimeoutTest)
   pthread_join(th, NULL);
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
 
   // Verify
   DiagStatus status;
@@ -734,7 +713,7 @@ TEST_F(CPUMonitorTestSuite, throttlingRecvNoDataTest)
   ThreadTestMode mode = RecvNoData;
   pthread_create(&th, nullptr, msr_reader, &mode);
   // Wait for thread started
-  rclcpp::WallRate(10).sleep();
+  ros::WallDuration(0.1).sleep();
 
   // Publish topic
   monitor_->update();
@@ -742,8 +721,8 @@ TEST_F(CPUMonitorTestSuite, throttlingRecvNoDataTest)
   pthread_join(th, NULL);
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
 
   // Verify
   DiagStatus status;
@@ -761,7 +740,7 @@ TEST_F(CPUMonitorTestSuite, throttlingFormatErrorTest)
   ThreadTestMode mode = FormatError;
   pthread_create(&th, nullptr, msr_reader, &mode);
   // Wait for thread started
-  rclcpp::WallRate(10).sleep();
+  ros::WallDuration(0.1).sleep();
 
   // Publish topic
   monitor_->update();
@@ -769,8 +748,8 @@ TEST_F(CPUMonitorTestSuite, throttlingFormatErrorTest)
   pthread_join(th, NULL);
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
 
   // Verify
   DiagStatus status;
@@ -788,8 +767,8 @@ TEST_F(CPUMonitorTestSuite, throttlingConnectErrorTest)
   monitor_->update();
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
 
   // Verify
   DiagStatus status;
@@ -807,8 +786,8 @@ TEST_F(CPUMonitorTestSuite, freqTest)
   monitor_->update();
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
 
   // Verify
   DiagStatus status;
@@ -825,8 +804,8 @@ TEST_F(CPUMonitorTestSuite, freqFrequencyFilesNotFoundTest)
   monitor_->update();
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
 
   // Verify
   DiagStatus status;
@@ -848,8 +827,9 @@ TEST_F(CPUMonitorTestSuite, usageMpstatErrorTest)
   monitor_->update();
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
+
   // Verify
   DiagStatus status;
   std::string value;
@@ -872,8 +852,8 @@ TEST_F(CPUMonitorTestSuite, usageMpstatExceptionTest)
   monitor_->update();
 
   // Give time to publish
-  rclcpp::WallRate(2).sleep();
-  rclcpp::spin_some(monitor_->get_node_base_interface());
+  ros::WallDuration(0.5).sleep();
+  ros::spinOnce();
 
   // Verify
   DiagStatus status;
@@ -890,18 +870,15 @@ class DummyCPUMonitor : public CPUMonitorBase
   friend class CPUMonitorTestSuite;
 
 public:
-  DummyCPUMonitor(const std::string & node_name, const rclcpp::NodeOptions & options)
-  : CPUMonitorBase(node_name, options)
+  DummyCPUMonitor(const ros::NodeHandle & nh, const ros::NodeHandle & pnh) : CPUMonitorBase(nh, pnh)
   {
   }
-  void update() { updater_.force_update(); }
+  void update(void) { updater_.force_update(); }
 };
 
 TEST_F(CPUMonitorTestSuite, dummyCPUMonitorTest)
 {
-  rclcpp::NodeOptions options;
-  std::unique_ptr<DummyCPUMonitor> monitor =
-    std::make_unique<DummyCPUMonitor>("dummy_cpu_monitor", options);
+  std::unique_ptr<DummyCPUMonitor> monitor = std::make_unique<DummyCPUMonitor>(nh_, pnh_);
   monitor->getTempNames();
   monitor->getFreqNames();
   // Publish topic
@@ -912,6 +889,7 @@ int main(int argc, char ** argv)
 {
   argv_ = argv;
   testing::InitGoogleTest(&argc, argv);
+  ros::init(argc, argv, "CPUMonitorTestNode");
 
   return RUN_ALL_TESTS();
 }
